test(web): split alpha RFQ e2e journeys
